Added a custom part template

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -97,6 +97,9 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::find($id);
+        if($category->posts->count()){
+            return redirect()->route('categories.index')->with('error','Ошибка! У категории есть запеси.');
+        }
         $category->delete();
         return redirect()->route('categories.index')->with('success','Категория удалена');
     }
