+++ app/Http/Controllers/Admin/TagController.php
@@ -96,7 +96,10 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
+
+
         $tag = Tag::find($id);
+        if($tag->posts->count()) return redirect()->route('tag.index')->with('error','Ошибка! У тегов есть запеси.');
         $tag->delete();
         return redirect()->route('tags.index')->with('success','Тег удален');
     }
